fix: 9 errores de compilacion CS0103/CS0104/CS0136/CS0191/CS0234/CS0266
GeneradorTerrenoUltraPreciso:
  + using System  (Buffer.BlockCopy)
  OX/OZ -> literales 1918f/8570f (no estaban en scope)

ExtensionesSeguras:
  System.MissingComponentException -> UnityEngine.MissingComponentException

AlsasuaTreeStreamer:
  _instancias = list  (readonly) -> _instancias.RemoveAll(null)
  Random.Range -> UnityEngine.Random.Range (ambiguedad Unity.Mathematics)

SistemaManifestacion:
  Random.Range -> UnityEngine.Random.Range (ambiguedad Unity.Mathematics)

GeneradorGeometriaPrecisa:
  var matReal (inner scope) -> var matLIDAR  (CS0136 conflicto de nombre)

IntegradorAssets:
  veg.meshesArbol (eliminado en rewrite) -> log directo, sin acceso al campo

SistemaSueloAAA:
  _terrain -> terrain (variable local ya declarada en la misma funcion)

ProcesadorNubePuntos:
  float lon = ... + LON0 -> + (float)LON0  (const double a float)
  float lat = ... + LAT0 -> + (float)LAT0

+++ Assets/Scripts/ProcesadorNubePuntos.cs
@@ -355,13 +355,13 @@ public class ProcesadorNubePuntos : MonoBehaviour
 
     int UnityXToCol(float x)
     {
-        float lon = (x - OX) / M_PER_DEG_LON + LON0;
+        float lon = (x - OX) / M_PER_DEG_LON + (float)LON0;
         return Mathf.RoundToInt((lon - _dsmXllCorner) / _dsmCellSize);
     }
 
     int UnityZToRow(float z)
     {
-        float lat = (z - OZ) / M_PER_DEG_LAT + LAT0;
+        float lat = (z - OZ) / M_PER_DEG_LAT + (float)LAT0;
         return _dsmNrows - 1 - Mathf.RoundToInt((lat - _dsmYllCorner) / _dsmCellSize);
     }
